+++ Plugin/Non_Modal_Dialogue_Interface.h
@@ -17,10 +17,6 @@
 
 #include "Min_Win_Defs.h"
 
-// This doesn't actually provide anything useful, but IWYU can't currently
-// work out that functions declared as override don't need to a separate header
-#include <intsafe.h>
-
 // Forward references
 class Plugin;
 
@@ -29,7 +25,7 @@ class Plugin;
  */
 class Non_Modal_Dialogue_Interface : public Non_Modal_Dialogue_Base
 {
-    typedef Non_Modal_Dialogue_Base Super;
+    using Super = Non_Modal_Dialogue_Base;
 
   public:
     /** Create a docking dialogue.
@@ -56,5 +52,5 @@ class Non_Modal_Dialogue_Interface : public Non_Modal_Dialogue_Base
      */
     Message_Return on_unhandled_non_modal_dialogue_message(
         UINT message, WPARAM wParam, LPARAM lParam
-    ) noexcept override final;
+    ) noexcept final;
 };